% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_plot.R
\encoding{UTF-8}
\name{ce_plot}
\alias{ce_plot}
\title{ce_plot}
\usage{
ce_plot(data = NULL, c_source = NULL, location_g = NULL, type = NULL)
}
\arguments{
\item{data}{list. A list storing matrices containing the mean and standard
deviation of the climate and/or elevation data.}

\item{c_source}{Character. Supply either \code{"CHELSA"} or
\code{"WorldClim"} argument.}

\item{location_g}{Character. Corresponding to the area for which the diagram
will be produced.}

\item{type}{Character. To produce Walter-Lieth ("WL") or Holdridge ("H")
diagrams}
}
\value{
Returns a 'ggplot2' family of plot. This function uses the climaemet
package to create the Walter and Lieth (XXXX) climatic diagram and the
macroBiome and Ternary packages to create a Holdridge simplex plot.
}
\description{
Creates a graph using the climate and elevation data which has
been extracted for a given \code{location}. It accepts the data formatted
from the \code{ce_extract} function.
}
\examples{
\dontrun{

# Extraction time will depend on the size of the polygon or point file.
# Import the Sibillini National Park Boundary
data(Sibillini_py)
# Run the download function
ce_extract(
dir.clim = "../WorkingDirectory/CHELSA",
dir.elev = "../WorkingDirectory/elev",
location = Sibillini_py,
location_g = "Position"
c_source = "CHELSA")

data <- ce_extract(dir.clim = dir.clim, dir.elev = dir.elev,
location = Sibillini_py, location_g = location_g,
c_source = "CHELSA", var = "ALL")

ce_plot(data = data, c_source = "CHELSA", location_g = "High",
type = 'WL')

ce_plot(data = data, c_source = "CHELSA", location_g = "High",
type = 'H')
}

}
\references{
{ Holdridge (1947), Determination of world plant formations from
simple climatic data. Science, 105:367–368.
\doi{10.1126/science.105.2727.367}

Holdridge (1967), Life Zone Ecology. Tropical Science Center, San José.

Pizarro, M, Hernangómez, D. & Fernández-Avilés G. (2023).
climaemet: Climate AEMET Tools. Comprehensive R Archive Network.
\doi{10.5281/zenodo.5205573}

Szelepcsényi, Z. (2023) macroBiome: A Tool for Mapping the
Distribution of the Biomes and Bioclimate. Comprehensive R Archive Network.
\doi{10.5281/zenodo.7633367}

Smith, M.R (2017). Ternary: An R Package for Creating Ternary Plots.
Comprehensive R Archive Network. \doi{10.5281/zenodo.1068996}

Walter, H. & Lieth, H. (1960). Klimadiagramm Weltatlas. G. Fischer, Jena.

}
}
\seealso{
Download climate data: \code{\link[=ce_download]{ce_download()}}
}
\author{
James L. Tsakalos
}
