% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev.R
\name{elev}
\alias{elev}
\title{elev}
\usage{
elev(output_dir = NULL, location = NULL, source = "mapzen")
}
\arguments{
\item{output_dir}{Character (e.g., ".../Desktop/chelsa"). Pathway to where
the data will be stored.}

\item{location}{A \code{sp}, \code{sf} polygon or point object.}

\item{source}{Character. Elevation sourced from mapzen (default) or geodata.}
}
\value{
Returns one elevation tile (tiff). If elevation is sourced from geodata the
elevation is downloaded at a spatial resolution of 30 arc seconds (~1 km2).
If elevation data is sourced from mapzen then the product will be a mosaic.
Specfically, Mapzen’s product is unique in that it combines several sources
of digital elevation models, including SRTM, the ArcticDEM (covering all
areas north of 60°), EUDEM (digital elevation model over Europe; for review,
see Mouratidis & Ampatzidis, 2019), and others into a single product. The
resolution of this product was set to 7, corresponding to 611.5 m ground
resolution at 60° latitude 864.8 m at 45° and 1223 m at 0°.
}
\description{
\code{elev} downloads elevation data the Shuttle Radar Topography Mission
(SRTM) , specifically the hole-filled CGIAR-SRTM (90 m resolution) for
latitudes between -60 and 60 or Mapzen's synthesis digital elevation product.
}
\examples{
\dontrun{
# Start by loading Italy's Mount Sibillini National Park boundary
data("Sibillini_py")
# elevation will be saved in the output_dir (i.e. output directory)
elev(output_dir = "...Desktop/elev", location = Sibillini_py)
}
}
\references{
{ Hijmans, R.J., et al., (2023). geodata: Download Geographic
Data. R package version 0.5-8. https://CRAN.R-project.org/package=geodata

Hollister, J. W., et al. (2021). elevatr: Access Elevation Data from Various
APIs. R package version 0.4.2. https://CRAN.R-project.org/package=elevatr

Mouratidis, A., and Ampatzidis, D. (2019). European Digital Elevation Model
Validation against Extensive Global Navigation Satellite Systems Data and
Comparison with SRTM DEM and ASTER GDEM in Central Macedonia (Greece).
ISPRS International Journal of Geo-Information 8, 108.
DOI: 10.3390/ijgi8030108
}
}
\seealso{
\code{\link{ce_download.R}}
}
\author{
James L. Tsakalos
}
