% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_it_py.R
\docType{data}
\encoding{UTF-8}
\name{it_py}
\alias{it_py}
\title{Italian Biome polygon data}
\format{
{ An object of class \code{sf} (inherits from \code{tbl_df},
\code{tbl}, \code{data.frame}) with 328 rows and 2 columns.
}
}
\usage{
it_py
}
\description{
The package contains geospatial polygon data showing the Biomes of Italy.
}
\details{
The first column, \code{geometry}, refers to the X and Y coordinate
of each polygon spaced over Italy. The second column, \code{GB}, refers to
the Global Biome assigned to each coordinate. For more details on the
assignment of Biomes please refer to Mucina et al. (2022).
}
\examples{
data("it_py")
plot(it_py)
}
\references{
{ Mucina, L., Divíšek, J., Tsakalos, J.L. (2022).
Europe, Ecosystems of. Reference Module in Life Sciences.
\doi{10.1016/B978-0-12-822562-2.00059-1}
}
}
\keyword{datasets}
