% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_c.R
\encoding{UTF-8}
\name{plot_c}
\alias{plot_c}
\title{plot_c}
\usage{
plot_c(
  data,
  geo_id,
  p_cex = 1,
  nchr_main = 32,
  l_main = 0.1,
  l_prec = 0,
  l_month = -0.6,
  lwd_temp = 1.5,
  lwd_prec = 9,
  l_units = 0.1,
  l_tcols = c(14.5, 16.5, 18.5, 19.5, 22.5)
)
}
\arguments{
\item{geo_id}{Character. Corresponding to a specific feature contained in the
\code{"location"} argument (specifically in the \code{"location_g"} column).}

\item{p_cex}{Numeric. Text size for entire plot.}

\item{nchr_main}{Numeric. Controls the number of characters of the main title
before wrapping to a new line.}

\item{l_main}{Numeric. Line position of the main title.}

\item{l_prec}{Numeric. Line position of the y axis (precipitation).}

\item{l_month}{Numeric. Line position of the x axis (months).}

\item{lwd_temp}{Numeric. Line width of temperature.}

\item{lwd_prec}{Numeric. Line width of precipitation.}

\item{l_units}{Numeric. Line position of precipitation and temperature units.}

\item{l_tcols}{List. Line position of the table columns. Must be length 5
corresponding to the position (left to right) for each column.}

\item{output_data_param}{List. Containing climate, elevation and latitude
data sets. Structured by \code{ce_extract()}.}
}
\value{
Returns a base R family of plot. This function uses the
\pkg{dismo} package to calculate isothermality (ISO),
temperature seasonality (TS) and precipitation seasonality (PS). It also uses
the \pkg{macroBiome} package to calculate mean annual biotemperature (BioT)
and the potential evapotranspiration ratio (PET).
}
\description{
Creates a graph using the climate and elevation data which has
been extracted for a given \code{location}. It accepts the data formatted
from the \code{ce_extract} function.
}
\examples{
{

# Step 1. Import the Italian Biome polygon data
# Step 2. Run the download function
# Step 3. Run the extract function
#* See ce_download & ce_extract documentation

# Steps 1, 2 & 3 can be skipped by loading the extracted data (it_data)
data("it_data")

# Step 4. Visualise the climatic envelope using our custom diagram

plot_c(data = it_data, geo_id = "MED")

}

}
\references{
{ Pizarro, M, Hernangómez, D. &
Fernández-Avilés G. (2023). climaemet: Climate AEMET Tools.
Comprehensive R Archive Network. \doi{10.5281/zenodo.5205573}

Walter, H. & Lieth, H. (1960). Klimadiagramm Weltatlas. G. Fischer, Jena.

}
}
\seealso{
Download climate data: \code{\link[=ce_download]{ce_download()}}
}
\author{
James L. Tsakalos
}
