% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_h.R
\encoding{UTF-8}
\name{plot_h}
\alias{plot_h}
\title{plot_h}
\usage{
plot_h(data, location_g, col = "red", pch = 19, ...)
}
\arguments{
\item{data}{List. A list storing matrices containing the mean and standard
deviation of the climate and/or elevation data.}

\item{location_g}{Character. Informs how the zonal statistics are exported.
Must correspond to a column of the \code{"location"} argument. If NULL,
the zonal statistics are calculated for all features of \code{"location"}.}

\item{col, pch, \dots}{Arguments to control point styling in
\code{HoldridgePoints()}.}
}
\value{
Returns a 'ggplot2' family of plot. This function uses the
\pkg{macroBiome} and \pkg{Ternary} packages to create a Holdridge simplex
plot.
}
\description{
Creates a graph using the climate and elevation data which has
been extracted for a given \code{location}. It accepts the data formatted
from the \code{ce_extract} function.
}
\examples{
{

# Step 1. Import the Italian Biome polygon data
# Step 2. Run the download function
# Step 3. Run the extract function
#* See ce_download & ce_extract documentation

# Steps 1, 2 & 3 can be skipped by loading the extracted data (it_data)
data("it_data")

# Step 4. Visualise the climatic envelope using a Holdridge diagram

plot_h(data = it_data, location_g = "MED")

}

}
\references{
{ Holdridge (1947), Determination of world plant formations from
simple climatic data. Science, 105:367–368.
\doi{10.1126/science.105.2727.367}

Holdridge (1967). Life Zone Ecology. Tropical Science Center,
San José.

Szelepcsényi, Z. (2023) macroBiome: A Tool for Mapping the
Distribution of the Biomes and Bioclimate. Comprehensive R Archive Network.
\doi{10.5281/zenodo.7633367}

Smith, M.R (2017). Ternary: An R Package for Creating Ternary Plots.
Comprehensive R Archive Network. \doi{10.5281/zenodo.1068996}

}
}
\seealso{
Download climate data: \code{\link[=ce_download]{ce_download()}}
}
\author{
James L. Tsakalos
}
