% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wl.R
\encoding{UTF-8}
\name{plot_wl}
\alias{plot_wl}
\title{plot_wl}
\usage{
plot_wl(data, location_g, ...)
}
\arguments{
\item{data}{List. A list storing matrices containing the mean and standard
deviation of the climate and/or elevation data.}

\item{location_g}{Character. Informs how the zonal statistics are exported.
Must correspond to a column of the \code{"location"} argument. If NULL,
the zonal statistics are calculated for all features of \code{"location"}.}

\item{\dots}{Arguments to control styling in
\code{ggclimat_walter_lieth()}.}
}
\value{
Returns a 'ggplot2' family of plot. This function uses the
\pkg{climaemet} package to create the Walter and Lieth (1960) climatic
diagram.
}
\description{
Creates a graph using the climate and elevation data which has
been extracted for a given \code{location}. It accepts the data formatted
from the \code{ce_extract} function.
}
\examples{
{

# Step 1. Import the Sibillini National Park Boundary
# Step 2. Run the download function
# Step 3. Run the extract function
#* See ce_download & ce_extract documentation

# Steps 1, 2 & 3 can be skipped by loading the extracted data
data(s_data)

# Step 4. Visualise the climatic envelope using a Walter-Lieth diagram

plot_wl(data = s_data, location_g = "High")

}

}
\references{
{ Pizarro, M, Hernangómez, D. &
Fernández-Avilés G. (2023). climaemet: Climate AEMET Tools.
Comprehensive R Archive Network. \doi{10.5281/zenodo.5205573}

Walter, H. & Lieth, H. (1960). Klimadiagramm Weltatlas. G. Fischer, Jena.

}
}
\seealso{
Download climate data: \code{\link[=ce_download]{ce_download()}}
}
\author{
James L. Tsakalos
}
