% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sibillini}
\alias{Sibillini}
\alias{Sibillini_pt}
\alias{Sibillini_py}
\alias{s_data}
\title{Sibillini data}
\format{
An object of class \code{SpatialPointsDataFrame} with 100 rows and 3 columns.

An object of class \code{SpatialPolygonsDataFrame} with 4 rows and 2 columns.

An object of class \code{list} of length 12.
}
\usage{
Sibillini_pt

Sibillini_py

s_data
}
\description{
Geospatial polygon and point data sets capturing Italy's Monti
'Sibillini' National Park.
}
\details{
The polygon data set ('Sibillini_py') contains two polygons dividing
the park into regions of 'high' and 'low' altitude.
The point data set ('Sibillini_pt') contains 100 points distributed randomly
within the 'high' and 'low' sections.

TODO: Complete description of \code{Sibillini_pt} and \code{Sibillini_py}
}
\examples{
data("Sibillini_pt")
#TODO: Do something with this data

}
\keyword{datasets}
