% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibillini_pt.R
\docType{data}
\encoding{UTF-8}
\name{Sibillini_pt}
\alias{Sibillini_pt}
\title{Sibillini point data}
\format{
{ An object of class \code{sf} (inherits from \code{tbl_df},
\code{tbl}, \code{data.frame}) with 9963 rows and 2 columns.
}
}
\usage{
Sibillini_pt
}
\description{
The package contains geospatial point data capturing Italy's Monti
'Sibillini' National Park.
}
\details{
Contains 100 points distributed randomly within the 'high' and
'low' sections.
}
\examples{
data("Sibillini_pt", package = "climenv")
terra::plot(Sibillini_pt)
}
\references{
{ Mucina, L., Divíšek, J., Tsakalos, J.L. (2022).
Europe, Ecosystems of. Reference Module in Life Sciences.
\doi{10.1016/B978-0-12-822562-2.00059-1}
}
}
\keyword{datasets}
