% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibillini_py.R
\docType{data}
\encoding{UTF-8}
\name{Sibillini_py}
\alias{Sibillini_py}
\title{Sibillini polygon data}
\format{
{ An object of class \code{sf} (inherits from \code{tbl_df},
\code{tbl}, \code{data.frame}) with 328 rows and 2 columns.
}
}
\usage{
Sibillini_py
}
\description{
The package contains geospatial point data capturing Italy's Monti
'Sibillini' National Park.
}
\details{
Contains two polygons dividing the park into regions of 'high' and
'low' altitude.
}
\examples{
data("Sibillini_py", package = "climenv")
terra::plot(Sibillini_py)
}
\references{
{ Mucina, L., Divíšek, J., Tsakalos, J.L. (2022).
Europe, Ecosystems of. Reference Module in Life Sciences.
\doi{10.1016/B978-0-12-822562-2.00059-1}
}
}
\keyword{datasets}
