% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srtm.tiles}
\alias{srtm.tiles}
\title{STRM tile data}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 872 rows and 1 columns.
}
\usage{
srtm.tiles
}
\description{
TODO: Description of \code{srtm.tiles}
}
\keyword{datasets}
