% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_srtm_tiles.R
\docType{data}
\encoding{UTF-8}
\name{srtm_tiles}
\alias{srtm_tiles}
\title{SRTM tiles data}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 872 rows
and 1 column.
}
\usage{
srtm_tiles
}
\description{
Polygons defining coordinates of Shuttle Radar Topography Mission data
tiles.
}
\details{
Contains tiles to assist downloading from 'geodata'.
}
\examples{
data("srtm_tiles", package = "climenv")
head(srtm_tiles)
}
\keyword{datasets}
