% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_extract.R
\name{ce_extract}
\alias{ce_extract}
\title{ce_extract}
\usage{
ce_extract(
  dir_clim = NULL,
  dir_elev = NULL,
  location = NULL,
  location_g = NULL,
  c_source = NULL,
  var = "ALL"
)
}
\arguments{
\item{dir_clim}{Character. Directory containing the climate data. Must
feature prec, tmax, tmean and tmin subfolders.}

\item{dir_elev}{Character. Directory containing the elevation data.}

\item{location}{A \code{sp}, \code{sf} polygon or point object.}

\item{location_g}{Character. Informs how the zonal statistics are exported.
Should correspond to a column of the \code{location} argument. If NULL,
the zonal statistics are calculated for all features of \code{location}.}

\item{c_source}{Character. Supply either \code{"CHELSA"} or
\code{"WorldClim"} argument.}

\item{var}{Character. If supplied will export a subset of the climate or
elevation data.}
}
\value{
Returns a list storing matrices containing the mean and standard deviation
of the climate and/or elevation data. Each column represents a month, each
row represents a feature of the \code{location} \code{sp}, \code{sf} polygon
or point object. Values returned are either degrees Celsius for (tmax, tmean,
tmin) or mm (prec).
}
\description{
\code{ce_extract} brings together the \code{worldclim},
\code{chelsa} and \code{elev} functions into one location to streamline data
downloading.
}
\examples{
\dontrun{

# Extraction time will depend on the size of the polygon or point file.
# Import the Sibillini National Park Boundary
data(Sibillini_py)
# Run the download function
ce_extract(
dir_clim = "../WorkingDirectory/CHELSA",
dir_elev = "../WorkingDirectory/elev",
location = Sibillini_py,
location_g = "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
c_source = "CHELSA")

See also PLOTTTTTTTIN XXXXX functions.
}
}
\seealso{
\code{\link{ce_download.R}}
}
\author{
James L. Tsakalos
}
