% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_extract.R
\name{ce_extract}
\alias{ce_extract}
\title{ce_extract}
\usage{
ce_extract(
  dir_clim = NULL,
  dir_elev = NULL,
  location = NULL,
  location_g = NULL,
  c_source = "WorldClim",
  var = "ALL"
)
}
\arguments{
\item{dir_clim}{Character. Directory containing the climate data. Must
feature prec, tmax, tmean and tmin subfolders. Within each subfolder there
should be a 12 raster files (e.g. \code{"prec01.tif"}) corresponding to the
monthly climate data and with a spatial extent that encompasses the area
covered by the polygon or point object.}

\item{dir_elev}{Character. Directory containing the elevation data. Within
the folder there should be a raster file ("strm.tif") with a spatial extent
that encompasses the area covered by the polygon or point object.}

\item{location}{A \code{"sp"}, \code{"sf"} polygon or point object.}

\item{location_g}{Character. Informs how the zonal statistics are exported.
Must correspond to a column of the \code{"location"} argument. If NULL,
the zonal statistics are calculated for all features of \code{"location"}.}

\item{c_source}{Character. Climatic data source, must be either \code{"CHELSA"} or
\code{"WorldClim"} (default). MS: We should support partial, case-insensitive
matching.}

\item{var}{Character. If supplied will download a subset of the climate data.
Must be one of \code{"ALL"} (default), \code{"prec"}, \code{"tmax"}, \code{"tmin"} or \code{"tmean"}
to download the corresponding climate data.}
}
\value{
Returns a list storing matrices containing the mean and standard deviation
of the climate and/or elevation data. Each column represents a month, each
row represents a feature of the \code{location} \code{sp}, \code{sf} polygon
or point object. Values returned are either degrees Celsius for (tmax, tmean,
tmin) or mm (prec).
}
\description{
Extracts climate and/or elevation data for generated over an area or at fixed
point/s.
}
\examples{
\dontrun{

# Extraction time will depend on the size of the polygon or point file.
# Import the Sibillini National Park Boundary
data("Sibillini_py")
# Run the download function
ce_extract(
  dir_clim = "../WorkingDirectory/CHELSA",
  dir_elev = "../WorkingDirectory/elev",
  location = Sibillini_py,
  location_g = "Position",
  c_source = "CHELSA"
)

}
}
\seealso{
The downloading (\code{\link[=ce_download]{ce_download()}}), and the plotting
(\code{\link[=plot_h]{plot_h()}} & \code{\link[=plot_wl]{plot_wl()}}) functions.
}
\author{
James L. Tsakalos
}
