% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chelsa.R
\name{chelsa}
\alias{chelsa}
\title{chelsa}
\usage{
chelsa(output_dir = NULL, mode = "wb", quiet = FALSE, var = NULL)
}
\arguments{
\item{output_dir}{Character (e.g., ".../Desktop/chelsa"). Pathway to where
the data will be stored.}

\item{mode}{Character. See documentation in /code{download.file()}.}

\item{quiet}{Logical. If TRUE, suppress status messages (if any), and the
progress bar.}

\item{var}{Character specifying which variables to extract.
Options are "prec", "tmax", "tmin", "tmean".}
}
\value{
Returns four subfolders named prec, tmax, tmin and tmean. Each folder
contains 12 GeoTiff (.tif) files, one for each month of the year
(January is 1; December is 12) for the time period 1981–2010. Each of the
files are downloaded at a spatial resolution of 30 arc seconds (~1 km sq.).
The precipitation folder contains average monthly precipitation (mm). The
tmax folder contains maximum monthly temperature. The tmin folder contains
minimum monthly temperature. The tmean folder contains the average monthly
temperature. The units for all temperature derived layers, as provided by
CHELSA Version 1.2, are multiplied by a factor of 10; hence, the unit of
measure for temperature is in °C X 10.. Specifically, monthly values of
minimum, maximum and mean.
}
\description{
\code{chelsa} downloads the CHELSA (Climatologies at high resolution for the
earth’s land surface areas) V2.1 data of downscaled temperature and
precipitation to a resolution of 30 arc sections.
}
\examples{
\dontrun{
chelsa(output_dir = "...Desktop/Chelsa", quiet = FALSE)
}
}
\references{
{ Karger, D.N., et al. (2016). CHELSA climatologies at high
resolution for the earth’s land surface areas (Version 1.0). World Data
Center for Climate. DOI: 10.1594/WDCC/CHELSA_v1

Karger, D.N., et al. (2016). CHELSA climatologies at high resolution for the
earth’s land surface areas (Version 1.1). World Data Center for Climate. DOI:
10.1594/WDCC/CHELSA_v1_1

Karger, D.N., et al. (2018). Data from: Climatologies at high
resolution for the earth’s land surface areas. EnviDat.
DOI: 10.16904/envidat.228.v2.1
}
}
\seealso{
Download climate data: \code{\link[=ce_download]{ce_download()}}
}
\author{
James L. Tsakalos
}
