% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sibillini}
\alias{Sibillini}
\alias{Sibillini_pt}
\alias{Sibillini_py}
\title{Sibillini data}
\format{
An object of class \code{SpatialPointsDataFrame} with 100 rows and 3 columns.

An object of class \code{SpatialPolygonsDataFrame} with 4 rows and 2 columns.
}
\usage{
Sibillini_pt

Sibillini_py
}
\description{
TODO: Description of \code{Sibillini_pt} and \code{Sibillini_py}
}
\keyword{datasets}
