% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRM_tiles_data.R
\docType{data}
\encoding{UTF-8}
\name{srtm_tiles}
\alias{srtm_tiles}
\title{STRM tiles data}
\format{
{ An object of class \code{SpatialPolygonsDataFrame} with 872 rows
and 1 columns.}
}
\usage{
srtm_tiles
}
\description{
SRTM tiles data.
}
\details{
Contains tiles to assist downloading from 'geodata'.
}
\examples{
data("srtm_tiles", package = "climenv")
head(srtm_tiles)
}
\references{
{ Mucina, L., Div&scaronek, J., Tsakalos, J.L. (2022).
Europe, Ecosystems of. Reference Module in Life Sciences.
\doi{10.1016/B978-0-12-822562-2.00059-1}
}
}
\keyword{datasets}
