% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{worldclim}
\alias{worldclim}
\title{Download WorldClim climate data}
\usage{
worldclim(output_dir, location, mode = "wb", var = "all", ...)
}
\arguments{
\item{output_dir}{Character (e.g., \code{"../Desktop/chelsa"}). Pathway to where
the data will be stored.}

\item{location}{A \code{"sp"}, \code{"sf"} polygon or point object.}

\item{var, mode, \dots}{Arguments to control a download from the Internet
\code{download.file()}.}
}
\value{
Creates four subfolders named prec, tmax, tmin and tmean. Each folder
contains 12 GeoTiff (.tif) files, one for each month of the year for the time
period 1970–2000. Each of the files are downloaded at a spatial
resolution of 0.5 minutes of a degree. The precipitation folder contains
average monthly precipitation (mm). The tmax folder contains maximum monthly
temperature. The tmin folder contains minimum monthly temperature. The tmean
folder contains the average monthly temperature. The unit of measure for
temperature is in °C. This function uses the \pkg{geodata} to download
the worldclim tiles.
}
\description{
\code{worldclim()} downloads the WorldClim V2.1 climate data for 1970–2000.
This includes monthly climate data for minimum, mean, and maximum temperature
and precipitation at a resolution of 0.5 minutes of a degree.
}
\examples{
\dontrun{
# Download time will depend on the size of the area you wish to access
# climate data for and your internet connection speed.

# Import the Italian Biome data set
data("it_py", package = "climenv")

# Download the WorldClim data
worldclim(
  output_dir = "...Desktop/worldclim",
  location = "it_py"
)

}
}
\references{
{ Fick, S.E. & Hijmans, R.J. (2017). WorldClim 2: new 1km
spatial resolution climate surfaces for global land areas. International
Journal of Climatology. 37, 4302–4315. \doi{10.1002/joc.5086}

}
}
\seealso{
Downloading from CHELSA \code{\link[=chelsa]{chelsa()}} or a more convenient
function for downloading other climate and elevation data \code{\link[=ce_download]{ce_download()}}.
}
\author{
James L. Tsakalos
}
